♻️ refactor [#59] : review api refactoring * QClass 새로 생성 * 패키지 구조 변경
