+++ build.gradle
@@ -27,8 +27,9 @@ dependencies {
     // QueryDSL
     implementation 'com.querydsl:querydsl-jpa:5.1.0:jakarta'
     annotationProcessor 'com.querydsl:querydsl-apt:5.1.0:jakarta'
-    annotationProcessor 'jakarta.annotation:jakarta.annotation-api'
+
     annotationProcessor 'jakarta.persistence:jakarta.persistence-api'
+    annotationProcessor 'jakarta.annotation:jakarta.annotation-api'
 
     // dotenv
     implementation 'io.github.cdimascio:java-dotenv:+'
@@ -81,3 +82,4 @@ dependencies {
 tasks.named('test') {
     useJUnitPlatform()
 }
+
