+++ src/main/java/com/sparta/spring_deep/_delivery/admin/review/ReviewAdminRepositoryCustomImpl.java
@@ -4,7 +4,7 @@ import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.types.Projections;
 import com.querydsl.core.types.dsl.DateTimePath;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import com.sparta.spring_deep._delivery.domain.review.QReview;
+import com.sparta.spring_deep._delivery.domain.review.model.QReview;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
