+++ src/main/java/com/sparta/spring_deep/_delivery/admin/review/ReviewAdminService.java
@@ -2,8 +2,8 @@ package com.sparta.spring_deep._delivery.admin.review;
 
 import com.sparta.spring_deep._delivery.domain.order.Order;
 import com.sparta.spring_deep._delivery.domain.order.OrderRepository;
-import com.sparta.spring_deep._delivery.domain.review.Review;
-import com.sparta.spring_deep._delivery.domain.review.ReviewResponseDto;
+import com.sparta.spring_deep._delivery.domain.review.model.Review;
+import com.sparta.spring_deep._delivery.domain.review.dto.ReviewResponseDto;
 import com.sparta.spring_deep._delivery.domain.user.entity.User;
 import com.sparta.spring_deep._delivery.exception.ResourceNotFoundException;
 import java.util.List;
