+++ src/main/java/com/sparta/spring_deep/_delivery/domain/order/OrderService.java
@@ -15,8 +15,8 @@ import com.sparta.spring_deep._delivery.domain.payment.PaymentResponseDto;
 import com.sparta.spring_deep._delivery.domain.payment.PaymentService;
 import com.sparta.spring_deep._delivery.domain.restaurant.Restaurant;
 import com.sparta.spring_deep._delivery.domain.restaurant.RestaurantRepository;
-import com.sparta.spring_deep._delivery.domain.review.Review;
-import com.sparta.spring_deep._delivery.domain.review.ReviewRepository;
+import com.sparta.spring_deep._delivery.domain.review.model.Review;
+import com.sparta.spring_deep._delivery.domain.review.repository.ReviewRepository;
 import com.sparta.spring_deep._delivery.domain.user.entity.User;
 import com.sparta.spring_deep._delivery.domain.user.entity.UserRole;
 import com.sparta.spring_deep._delivery.domain.user.repository.UserRepository;
