+++ src/test/java/com/sparta/spring_deep/_delivery/domain/review/ReviewIntegrationTest.java
@@ -24,6 +24,9 @@ import com.sparta.spring_deep._delivery.domain.restaurant.Restaurant;
 import com.sparta.spring_deep._delivery.domain.restaurant.RestaurantRepository;
 import com.sparta.spring_deep._delivery.domain.restaurant.restaurantAddress.RestaurantAddress;
 import com.sparta.spring_deep._delivery.domain.restaurant.restaurantAddress.RestaurantAddressRepository;
+import com.sparta.spring_deep._delivery.domain.review.dto.ReviewRequestDto;
+import com.sparta.spring_deep._delivery.domain.review.model.Review;
+import com.sparta.spring_deep._delivery.domain.review.repository.ReviewRepository;
 import com.sparta.spring_deep._delivery.domain.user.entity.User;
 import com.sparta.spring_deep._delivery.domain.user.entity.UserRole;
 import com.sparta.spring_deep._delivery.domain.user.repository.UserRepository;
